+++ sdk/lib/_internal/wasm/js_common/js_interop_patch.dart
@@ -226,6 +226,39 @@ extension FunctionToJSExportedDartFunction<T extends Function> on T {
   );
 }
 
+@patch
+extension FunctionToJSExportedDartFunctionVarArgs<
+  R extends JSAny?,
+  E extends JSAny?
+>
+    on R Function(JSArray<E>) {
+  @patch
+  JSExportedDartFunction<R Function(JSArray<E>)> get toJSVarArgs =>
+      js_helper.JS<JSFunction>('''(f) => {
+          const result = (...args) => f(args);
+          finalizeWrapper(f.dartFunction, result);
+          return result;
+        }''', this.toJS) as JSExportedDartFunction<R Function(JSArray<E>)>;
+}
+
+@patch
+extension FunctionToJSExportedDartFunctionCaptureThisVarArgs<
+  R extends JSAny?,
+  T extends JSObject,
+  E extends JSAny?
+>
+    on R Function(T, JSArray<E>) {
+  @patch
+  JSExportedDartFunction<R Function(T, JSArray<E>)>
+  get toJSCaptureThisVarArgs => js_helper.JS<JSFunction>('''(f) => {
+          const result = function(...args) {
+            return f(this, args);
+            finalizeWrapper(f.dartFunction, result);
+            return result;
+          };
+        }''', this.toJS) as JSExportedDartFunction<R Function(T, JSArray<E>)>;
+}
+
 // Embedded global property for wrapped Dart objects passed via JS interop.
 //
 // This is a Symbol so that different Dart applications don't share Dart